Move SonarAnalysisContextTest.ShouldAnalyzeGenerated
